Tested to confirm updates to classification model work with Keras models
