+++ credoai/utils/model_utils.py
@@ -1,8 +1,13 @@
 import warnings
+from credoai.utils import global_logger
 
 from sklearn.base import is_classifier, is_regressor
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.utils import multiclass
+
+import tensorflow as tf
+from tensorflow import keras
+from tensorflow.keras import layers
 
 
 def get_generic_classifier():
@@ -45,3 +50,45 @@ def get_default_metrics(model):
 
 def type_of_target(target):
     return multiclass.type_of_target(target) if target is not None else None
+
+
+#############################################
+# Validation Functions for Various Model Types
+#############################################
+def validate_sklearn_like(model_obj, model_info: dict):
+    pass
+
+
+def validate_keras_clf(model_obj, model_info: dict):
+    # This is how Keras checks sequential too: https://github.com/keras-team/keras/blob/master/keras/utils/layer_utils.py#L219
+    if not model_info["lib_name"] == "Sequential":
+        message = "Only Keras models with Sequential architecture are supported at this time. "
+        message += "Using Keras with other architechtures has undefined behavior."
+        global_logger.warning(message)
+
+    valid_final_layer = (
+        isinstance(model_obj.layers[-1], layers.Dense)
+        and model_obj.layers[-1].activation.__name__ == "softmax"
+    )
+    valid_final_layer = valid_final_layer or isinstance(
+        model_obj.layers[-1], layers.Softmax
+    )
+    if not valid_final_layer:
+        message = "Expected output layer to be either: tf.keras.layers.Softmax or "
+        message += "tf.keras.layers.Dense with softmax activation."
+        global_logger.warning(message)
+
+    if len(model_obj.layers[-1].output.shape) != 2:
+        message = "Expected 2D output shape for Keras.Sequetial model: (batch_size, n_classes) or (None, n_classes)"
+        global_logger.warning(message)
+
+    if model_obj.layers[-1].output.shape[0] is not None:
+        message = "Expected output shape of Keras model to have arbitrary length"
+        global_logger.warning(message)
+
+    if model_obj.layers[-1].output.shape[1] < 2:
+        message = "Expected classification output shape (batch_size, n_classes) or (None, n_classes). "
+        message += "Continuous output univariate regression not supported"
+        global_logger.warning(message)
+        # TODO Add support for model-imposed argmax layer
+        # https://stackoverflow.com/questions/56704669/keras-output-single-value-through-argmax
